rpi: rpi3: add canupdate.sh to allow installing compatible images
Prior to installing an image, confirm that the image is meant for an RPi,
and that the cpu is a BCM2837.

/proc/cpuinfo contains a "Revision:" field which is populated with
different information about the RPi device (manufacturer, cpu, etc.).

This field is only available through cpuinfo on 32-bit arm. 64-bit arm
devices need to access the field through the device tree, where it is stored
as hex. This is available in the same location on 32-bit arm as well.

Example field from cpuinfo: a22082
Example field from device-tree (with hexdump formatting): 00a22082

The significant digit for the cpu identification is the third (a2_2_082). For
a BCM2837, this digit will equal 2.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>